backend/requirements.txt
# PDF Komuta Merkezi - Backend Dependencies
# Profesyonel PDF Isleme ve Ceviri Sistemi

# Web Framework
flask>=3.0.0
flask-cors>=4.0.0
Werkzeug>=3.0.0
gunicorn>=21.0.0
python-dotenv>=1.0.0

# PDF Isleme
PyMuPDF>=1.23.0
pikepdf>=8.0.0

# PDF Olusturma (Turkce Font Destegi)
reportlab>=4.0.0
fpdf2>=2.7.0

# PDF -> Word
pdf2docx>=0.5.0
python-docx>=1.0.0

# PDF -> Excel
openpyxl>=3.1.0
pandas>=2.0.0

# AI Ceviri
google-generativeai>=0.5.0

# Goruntu Isleme
Pillow>=10.0.0

# Yardimci Araclar
numpy>=1.26.0
